drivers: wifi: siwx917: Split out NET_OFFLOAD code
siwx917_wifi.c is currently pretty large while a big part is in fact
only used to manage socket API. Split out the socket management and only
keep 802.11 management in siwx917_wifi.c.

This commit do not introduce any change in the code. The code is only
relocated.
